topics/vectorstore/rag-chroma-flyte/requirements.txt
flyte[tui]>=2.0
flyteplugins-vllm
gradio>=5.0.0
openai>=1.50.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
datasets>=3.0.0
python-dotenv
